initialized logic for members card
